requirements.txt
absl-py==0.1.10
backports.weakref==1.0.post1
bleach==1.5.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.4
Markdown==2.6.11
mock==2.0.0
numpy==1.14.0
pandas==0.22.0
pbr==3.1.1
Pillow==5.0.0
protobuf==3.5.1
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
scipy==1.0.0
six==1.11.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
Werkzeug==0.14.1
